Bugfix
The functioning of my brain appears to be up for debate

+++ programs/tangyongning-Malconv-90d940e/analysis.py
@@ -106,7 +106,7 @@ if __name__ == "__main__":
     if args[2] == "shapctrl":
         fraction = 0.5
         try:
-            fraction = int(args[1])
+            fraction = float(args[1])
         except ValueError:
             fraction = 0.5
